+++ src/validators/object/object.validator.js
@@ -17,6 +17,22 @@ const objectValidator = {
     const { getValidators } = require('../validators')
 
     const schemaObjProps = Object.keys(ctx.schema.props)
+    const valueProps = Object.keys(ctx.value)
+
+    if (ctx.schema.strict) {
+      const extraProps = valueProps.filter(prop => schemaObjProps.indexOf(prop) < 0)
+
+      if (extraProps.length > 0) {
+        ctx.registerErrors({
+          strict: {
+            message: getCurrentTranslation().validators.object.errors.strict(extraProps),
+            value: ctx.value,
+            extraProps
+          }
+        })
+        return 
+      }
+    }
 
     for (let propKey of schemaObjProps) {
       const propSchema = ctx.schema.props[propKey]